last code quality-check fixes
