+++ src/components/Settings/CapsulesSection/CapsulesSection.tsx
@@ -1,77 +1,17 @@
-import { useEffect } from 'react';
-import {
-	Box,
-	SimpleGrid,
-	Button,
-	Stack,
-	NumberInput,
-	Checkbox,
-	ColorInput,
-	Text,
-	Group,
-	Modal,
-	Divider,
-} from '@mantine/core';
-import { useForm } from '@mantine/form';
+import { Button, Text, Group, Modal, SimpleGrid, Box } from '@mantine/core';
 
 import { constants } from '@common/constants';
-import { CapsuleSettings } from '@extensionTypes/settingsValues';
 
 import { useExtensionSettings } from '@hooks/useExtensionSettings';
 import { useModal } from '@hooks/useModal';
 
-import { BookmarkCapsule } from '@components/BookmarkCapsule';
 import { showNotification } from '@mantine/notifications';
+import { CapsulesSettingsForm } from '@forms/CapsulesSettingsForm';
 
 export const CapsulesSection = () => {
-	const { extensionSettings, saveExtensionSettings } = useExtensionSettings();
+	const { saveExtensionSettings } = useExtensionSettings();
 
 	const resetCapsuleSettingsModal = useModal();
-
-	const { getInputProps, setValues, onSubmit, values, resetDirty, isDirty } =
-		useForm<CapsuleSettings>();
-
-	useEffect(() => {
-		if (extensionSettings) {
-			const capsuleSettings = Object.fromEntries(
-				Object.entries(extensionSettings).filter(([key]) => key.includes('capsule')),
-			) as CapsuleSettings;
-
-			setValues({
-				...capsuleSettings,
-				capsuleLabelColor:
-					capsuleSettings.capsuleLabelColor === null
-						? ''
-						: capsuleSettings.capsuleLabelColor,
-			});
-			resetDirty();
-		}
-	}, [extensionSettings]);
-
-	const handleSaveExtensionSettings = async (formValues: typeof values) => {
-		if (isDirty()) {
-			setTimeout(async () => {
-				try {
-					await saveExtensionSettings(formValues);
-
-					showNotification({
-						color: 'dark',
-						message: 'Settings for capsule view were successfully saved!',
-						autoClose: 3000,
-					});
-				} catch (error) {
-					showNotification({
-						color: 'red',
-						title: 'Settings could not be saved!',
-						message: 'Sorry, but something went wrong, please try again.',
-						autoClose: 5000,
-					});
-				}
-			}, 600);
-
-			resetDirty();
-		}
-	};
 
 	const handleResetSettings = async () => {
 		const defaultValues = Object.fromEntries(
@@ -100,90 +40,17 @@ export const CapsulesSection = () => {
 		}, 600);
 
 		resetCapsuleSettingsModal.close(); // hide modal
-		resetDirty();
 	};
 
 	return (
 		<>
-			<Box mb={32}>
-				<form onSubmit={onSubmit(handleSaveExtensionSettings)} noValidate>
-					<SimpleGrid cols={2} spacing={32} mb={8}>
-						<Text weight={700}>Customize:</Text>
-						<Text weight={700}>Preview:</Text>
-					</SimpleGrid>
+			<Box>
+				<SimpleGrid cols={2} spacing={32} mb={8}>
+					<Text weight={700}>Customize:</Text>
+					<Text weight={700}>Preview:</Text>
+				</SimpleGrid>
 
-					<SimpleGrid cols={2} spacing={16} mb={24}>
-						<Stack spacing={12} align="flex-start">
-							<NumberInput label="Capsule size" {...getInputProps('capsuleSize')} />
-
-							<NumberInput
-								label="Space between capsules"
-								{...getInputProps('capsuleSpacing')}
-							/>
-
-							<NumberInput
-								label="Favicon size"
-								{...getInputProps('capsuleIconSize')}
-							/>
-
-							<Group>
-								<NumberInput
-									label="Labels size"
-									{...getInputProps('capsuleLabelSize')}
-								/>
-
-								<ColorInput
-									{...getInputProps('capsuleLabelColor')}
-									format="hex"
-									label="Label color"
-								/>
-							</Group>
-
-							<Checkbox
-								label="Use bold text for labels"
-								{...getInputProps('capsuleLabelBold', {
-									type: 'checkbox',
-								})}
-							/>
-
-							<Checkbox
-								label="Use italic text for labels"
-								{...getInputProps('capsuleLabelItalic', {
-									type: 'checkbox',
-								})}
-							/>
-
-							<Checkbox
-								label="Hide labels for bookmarks"
-								{...getInputProps('capsuleHiddenName', {
-									type: 'checkbox',
-								})}
-							/>
-						</Stack>
-
-						<Group spacing={0}>
-							<Divider orientation="vertical" mr={16} />
-
-							<Group spacing={values.capsuleSpacing} mx="auto">
-								{constants.exampleBookmarks.map((bookmark) => (
-									<BookmarkCapsule
-										key={bookmark.id}
-										title={bookmark.name}
-										url={bookmark?.url}
-										settings={values}
-									/>
-								))}
-							</Group>
-						</Group>
-					</SimpleGrid>
-
-					<Group position="center" sx={{ width: '100%', marginLeft: '-28px' }} mt={0}>
-						<Button variant="outline" onClick={() => resetCapsuleSettingsModal.open()}>
-							Reset to default
-						</Button>
-						<Button type="submit">Save</Button>
-					</Group>
-				</form>
+				<CapsulesSettingsForm openResetModal={resetCapsuleSettingsModal.open} />
 			</Box>
 
 			<Modal
