+++ src/forms/CapsulesSettingsForm/index.ts
@@ -1,0 +1,1 @@
+export { CapsulesSettingsForm } from './CapsulesSettingsForm';
